Renamed GCFG to GlobalConfig for clarity

+++ JetAGN/JetAGNUnitTests/TestJetAGNResults.cpp
@@ -44,9 +44,9 @@ namespace JetAGNUnitTests
 		{
 			State model(config());
 			
-			Assert::AreEqual<size_t>(4, model.electron.ps[0].size(), L"energy dimension size", LINE_INFO());
-			Assert::AreEqual<size_t>(3, model.electron.ps[1].size(), L"energy dimension size", LINE_INFO());
-			Assert::AreEqual<size_t>(3, model.electron.ps[2].size(), L"energy dimension size", LINE_INFO());
+			Assert::AreEqual<int>(4, model.electron.ps[0].size(), L"energy dimension size", LINE_INFO());
+			Assert::AreEqual<int>(3, model.electron.ps[1].size(), L"energy dimension size", LINE_INFO());
+			Assert::AreEqual<int>(3, model.electron.ps[2].size(), L"energy dimension size", LINE_INFO());
 		}
 
 		TEST_METHOD(TestEnergyDimensionValues11)
